test(gates): add 268 audit gate workflows — feature, combo, error
H1-H4 of deep audit:
- 168 feature gates: all 5 verbs, 25 transforms, 12 builtins,
  for_each, bindings, artifacts, structured output, agent spawn
- 50 combo gates: cross-feature interactions
- 50 error gates: parse, DAG, binding, runtime errors
- Fix Vec<u32> -> Vec<u64> type mismatch in real_integration_test

All 168 feature + 50 combo gates pass nika check.

+++ tools/nika/tests/real_integration_test.rs
@@ -743,7 +743,7 @@ tasks:
     assert!(result.error.is_none());
 
     // Count turns with token info
-    let turn_tokens: Vec<u32> = result
+    let turn_tokens: Vec<u64> = result
         .events
         .iter()
         .filter_map(|e| {
